Code reorganization part #2
o renamed padgen.c / padgen.h to decryptor.c / decryptor.h
